requirements.txt
pandas>=2.0
numpy>=1.24
requests>=2.31
folium>=0.15
google-cloud-bigquery>=3.11
google-cloud-bigquery-storage>=2.22
pyarrow>=12.0
db-dtypes>=1.1
